Move the transformBody method into body.ts
