+++ content/webapp/services/prismic/transformers/body.ts
@@ -20,6 +20,7 @@ import {
   TitledTextList as TitledTextListSlice,
   GifVideoSlice,
   Discussion as DiscussionSlice,
+  Body,
 } from '../types/body';
 import { Props as TableProps } from '@weco/common/views/components/Table/Table';
 import { Props as ContactProps } from '@weco/common/views/components/Contact/Contact';
@@ -55,7 +56,7 @@ import {
 } from '.';
 import { transformTaslFromString } from '@weco/common/services/prismic/transformers';
 import { LinkField, RelationField, RichTextField } from '@prismicio/types';
-import { Weight } from '../../../types/generic-content-fields';
+import { BodyType, Weight } from '../../../types/generic-content-fields';
 import { Venue } from '@weco/common/model/opening-hours';
 import { transformCollectionVenue } from '@weco/common/services/prismic/transformers/collection-venues';
 import { GuidePrismicDocument } from '../types/guides';
@@ -95,7 +96,7 @@ type ParsedSlice<TypeName extends string, Value> = {
   value: Value;
 };
 
-export function transformStandfirstSlice(
+function transformStandfirstSlice(
   slice: StandfirstSlice
 ): ParsedSlice<'standfirst', RichTextField> {
   return {
@@ -105,7 +106,7 @@ export function transformStandfirstSlice(
   };
 }
 
-export function transformTextSlice(
+function transformTextSlice(
   slice: TextSlice
 ): ParsedSlice<'text', RichTextField> {
   return {
@@ -115,9 +116,7 @@ export function transformTextSlice(
   };
 }
 
-export function transformMapSlice(
-  slice: MapSlice
-): ParsedSlice<'map', MapProps> {
+function transformMapSlice(slice: MapSlice): ParsedSlice<'map', MapProps> {
   return {
     type: 'map',
     value: {
@@ -135,7 +134,7 @@ function transformTableCsv(tableData: string): string[][] {
     .map(row => row.split('|').map(cell => cell.trim()));
 }
 
-export function transformTableSlice(
+function transformTableSlice(
   slice: TableSlice
 ): ParsedSlice<'table', TableProps> {
   return {
@@ -150,7 +149,7 @@ export function transformTableSlice(
   };
 }
 
-export function transformMediaObjectListSlice(
+function transformMediaObjectListSlice(
   slice: MediaObjectListSlice
 ): ParsedSlice<'mediaObjectList', { items: MediaObjectType[] }> {
   return {
@@ -181,9 +180,7 @@ export function transformMediaObjectListSlice(
   };
 }
 
-export function transformTeamToContact(
-  team: TeamPrismicDocument
-): ContactProps {
+function transformTeamToContact(team: TeamPrismicDocument): ContactProps {
   const {
     data: { title, subtitle, email, phone },
   } = team;
@@ -196,7 +193,7 @@ export function transformTeamToContact(
   };
 }
 
-export function transformContactSlice(
+function transformContactSlice(
   slice: ContactSlice
 ): ParsedSlice<'contact', ContactProps> | undefined {
   return isFilledLinkToDocumentWithData(slice.primary.content)
@@ -207,7 +204,7 @@ export function transformContactSlice(
     : undefined;
 }
 
-export function transformEditorialImageSlice(
+function transformEditorialImageSlice(
   slice: EditorialImageSlice
 ): ParsedSlice<'picture', CaptionedImage> {
   return {
@@ -217,7 +214,7 @@ export function transformEditorialImageSlice(
   };
 }
 
-export function transformEditorialImageGallerySlice(
+function transformEditorialImageGallerySlice(
   slice: EditorialImageGallerySlice
 ): ParsedSlice<'imageGallery', ImageGalleryProps> {
   return {
@@ -230,7 +227,7 @@ export function transformEditorialImageGallerySlice(
   };
 }
 
-export function transformDeprecatedImageListSlice(
+function transformDeprecatedImageListSlice(
   slice: DeprecatedImageListSlice
 ): ParsedSlice<'deprecatedImageList', DeprecatedImageListProps> {
   return {
@@ -253,7 +250,7 @@ export function transformDeprecatedImageListSlice(
   };
 }
 
-export function transformGifVideoSlice(
+function transformGifVideoSlice(
   slice: GifVideoSlice
 ): ParsedSlice<'gifVideo', GifVideoProps> | undefined {
   const playbackRate = slice.primary.playbackRate
@@ -302,7 +299,7 @@ function transformTitledTextItem({
   };
 }
 
-export function transformTitledTextListSlice(
+function transformTitledTextListSlice(
   slice: TitledTextListSlice
 ): ParsedSlice<'titledTextList', TitledTextListProps> {
   return {
@@ -313,7 +310,7 @@ export function transformTitledTextListSlice(
   };
 }
 
-export function transformDiscussionSlice(
+function transformDiscussionSlice(
   slice: DiscussionSlice
 ): ParsedSlice<'discussion', DiscussionProps> {
   return {
@@ -325,7 +322,7 @@ export function transformDiscussionSlice(
   };
 }
 
-export function transformInfoBlockSlice(
+function transformInfoBlockSlice(
   slice: InfoBlockSlice
 ): ParsedSlice<'infoBlock', InfoBlockProps> {
   return {
@@ -339,7 +336,7 @@ export function transformInfoBlockSlice(
   };
 }
 
-export function transformIframeSlice(
+function transformIframeSlice(
   slice: IframeSlice
 ): ParsedSlice<'iframe', IframeProps> {
   return {
@@ -352,7 +349,7 @@ export function transformIframeSlice(
   };
 }
 
-export function transformQuoteSlice(
+function transformQuoteSlice(
   slice: QuoteSlice | QuoteV2Slice
 ): ParsedSlice<'quote', QuoteProps> {
   return {
@@ -367,7 +364,7 @@ export function transformQuoteSlice(
   };
 }
 
-export function transformTagListSlice(
+function transformTagListSlice(
   slice: TagListSlice
 ): ParsedSlice<'tagList', TagsGroupProps> {
   return {
@@ -385,7 +382,7 @@ export function transformTagListSlice(
   };
 }
 
-export function transformSearchResultsSlice(
+function transformSearchResultsSlice(
   slice: SearchResultsSlice
 ): ParsedSlice<'searchResults', AsyncSearchResultsProps> {
   return {
@@ -398,7 +395,7 @@ export function transformSearchResultsSlice(
   };
 }
 
-export function transformCollectionVenueSlice(
+function transformCollectionVenueSlice(
   slice: CollectionVenueSlice
 ):
   | ParsedSlice<
@@ -418,7 +415,7 @@ export function transformCollectionVenueSlice(
     : undefined;
 }
 
-export function transformEmbedSlice(
+function transformEmbedSlice(
   slice: EmbedSlice
 ): ParsedSlice<'videoEmbed' | 'soundcloudEmbed', EmbedProps> | undefined {
   const embed = slice.primary.embed;
@@ -484,7 +481,7 @@ export function transformEmbedSlice(
   }
 }
 
-export function transformContentListSlice(
+function transformContentListSlice(
   slice: ContentListSlice
 ): ParsedSlice<'contentList', SearchResultsProps> {
   type ContentListPrismicDocument =
@@ -534,4 +531,74 @@ export function transformContentListSlice(
         .filter(isNotUndefined),
     },
   };
+}
+
+export function transformBody(body: Body): BodyType {
+  return body
+    .map(slice => {
+      switch (slice.slice_type) {
+        case 'standfirst':
+          return transformStandfirstSlice(slice);
+
+        case 'text':
+          return transformTextSlice(slice);
+
+        case 'map':
+          return transformMapSlice(slice);
+
+        case 'editorialImage':
+          return transformEditorialImageSlice(slice);
+
+        case 'editorialImageGallery':
+          return transformEditorialImageGallerySlice(slice);
+
+        case 'titledTextList':
+          return transformTitledTextListSlice(slice);
+
+        case 'contentList':
+          return transformContentListSlice(slice);
+
+        case 'collectionVenue':
+          return transformCollectionVenueSlice(slice);
+
+        case 'searchResults':
+          return transformSearchResultsSlice(slice);
+
+        case 'quote':
+        case 'quoteV2':
+          return transformQuoteSlice(slice);
+
+        case 'iframe':
+          return transformIframeSlice(slice);
+
+        case 'gifVideo':
+          return transformGifVideoSlice(slice);
+
+        case 'contact':
+          return transformContactSlice(slice);
+
+        case 'embed':
+          return transformEmbedSlice(slice);
+
+        case 'table':
+          return transformTableSlice(slice);
+
+        case 'infoBlock':
+          return transformInfoBlockSlice(slice);
+
+        case 'discussion':
+          return transformDiscussionSlice(slice);
+
+        case 'tagList':
+          return transformTagListSlice(slice);
+
+        // Deprecated
+        case 'imageList':
+          return transformDeprecatedImageListSlice(slice);
+
+        case 'mediaObjectList':
+          return transformMediaObjectListSlice(slice);
+      }
+    })
+    .filter(isNotUndefined);
 }
